+++ scripts/macros/2024/feats/heavyArmorMaster.js
@@ -1,0 +1,4 @@
+export let heavyArmorMaster = {
+    name: 'Heavy Armor Master',
+    version: '1.2.36'
+};